Simplified volume serialization
Save the int values in memory and serialize those
Divide by 100.0 to get double values for sound backend
On deserialization, normalize int values to [0, 100]

+++ gframe/game_config.h
@@ -52,8 +52,8 @@ struct GameConfig
 	bool chkAnime = false;
 	bool enablemusic = false;
 	bool enablesound = true;
-	double musicVolume = 20;
-	double soundVolume = 20;
+	int musicVolume = 20;
+	int soundVolume = 20;
 	path_string skin = EPRO_TEXT("none");
 };
